attached_assets/Pasted-Uncaught-runtime-errors-ERROR-Element-type-is-invalid-expected-a-string-for-built-in-compone-1741504970430.txt

Uncaught runtime errors:
×
ERROR
Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `SideBarTabView`.
Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `SideBarTabView`.
    at createFiberFromTypeAndProps (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93195:21)
    at createFiberFromElement (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93216:19)
    at createChild (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80288:32)
    at reconcileChildrenArray (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80528:29)
    at reconcileChildFibers (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80870:20)
    at reconcileChildren (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:85290:32)
    at updateHostComponent (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:85941:7)
    at beginWork (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:87393:18)
    at HTMLUnknownElement.callCallback (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:72351:18)
    at Object.invokeGuardedCallbackDev (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:72395:20)
ERROR
Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `SideBarTabView`.
Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `SideBarTabView`.
    at createFiberFromTypeAndProps (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93195:21)
    at createFiberFromElement (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93216:19)
    at createChild (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80288:32)
    at reconcileChildrenArray (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80528:29)
    at reconcileChildFibers (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80870:20)
    at reconcileChildren (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:85290:32)
    at updateHostComponent (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:85941:7)
    at beginWork (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:87393:18)
    at HTMLUnknownElement.callCallback (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:72351:18)
    at Object.invokeGuardedCallbackDev (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:72395:20)
ERROR
Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `SideBarTabView`.
Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `SideBarTabView`.
    at createFiberFromTypeAndProps (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93195:21)
    at createFiberFromElement (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:93216:19)
    at createChild (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80288:32)
    at reconcileChildrenArray (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80528:29)
    at reconcileChildFibers (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:80870:20)
    at reconcileChildren (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:85290:32)
    at updateHostComponent (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:85941:7)
    at beginWork (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:87393:18)
    at beginWork$1 (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:92328:18)
    at performUnitOfWork (https://9a5d6691-1a6c-420a-b671-53dbf39b4aa6-00-1kz1sp6ggvcxo.kirk.replit.dev/react-rtc-demo/static/js/bundle.js:91598:16)
